+++ tests/testthat/test-raw-with-attributes.R
@@ -1,0 +1,13 @@
+
+test_that("writing a raw vector with attributes works", {
+
+  rv <- raw() |> 
+    set_endian('big') |>
+    write_uint32(1)
+
+  expect_true(!is.null(attributes(rv)))
+  
+  expect_no_error({  
+    write_raw(raw(), rv)
+  })
+})